Update mermaid to 8.13.2 and support background colors
As mermaid diagram with light themes are barely readable on dark backgrounds (and vice-versa),
this commit make it possible to force a background color by appending the following query parameter to any URL: `?bgColor=<color>`.
The color value is interpreted as an hexadecimal color code by default (`FF0000` for red), but you can also use named colors by prefixing the value with `!`.